Removed old firebase files from client/ init and instead initialised firebase for the whole monorepo. Reorganised file structure - placing OCR/ and algorithm/ inside server directory. Uninstalled packages from root package.json. Configured Google Cloud Run to run containerised server code. Now scrabblr is hosted on firebase and uses Cloud Run to execute server code running much faster than it did on render! phew
