+++ client/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "firebase": "^11.6.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
