+++ client/src/firebase.js
@@ -1,0 +1,14 @@
+import { initializeApp } from 'firebase/app';
+import { getFirestore } from 'firebase/firestore';
+
+const firebaseConfig = {
+	apiKey: 'YOUR_API_KEY',
+	authDomain: 'YOUR_PROJECT.firebaseapp.com',
+	projectId: 'YOUR_PROJECT',
+	storageBucket: 'YOUR_PROJECT.appspot.com',
+	messagingSenderId: 'YOUR_SENDER_ID',
+	appId: 'YOUR_APP_ID',
+};
+
+const app = initializeApp(firebaseConfig);
+export const db = getFirestore(app);
