+++ firebase.json
@@ -1,0 +1,19 @@
+{
+	"hosting": {
+		"public": "client/dist",
+		"ignore": ["firebase.json", "**/.*", "**/node_modules/**"],
+		"rewrites": [
+			{
+				"source": "/api/**",
+				"run": {
+					"serviceId": "backend-service",
+					"region": "australia-southeast1"
+				}
+			},
+			{
+				"source": "**",
+				"destination": "/index.html"
+			}
+		]
+	}
+}
