+++ package.json
@@ -16,10 +16,7 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
-		"@google-cloud/vision": "^4.3.2",
-		"dotenv": "^16.4.7",
-		"object-sizeof": "^2.6.5",
-		"word-list": "^4.1.0"
+		"dotenv": "^16.4.7"
 	},
 	"devDependencies": {
 		"jest": "^29.7.0"
