+++ server/OCR/util/createBoardState.js
