+++ server/OCR/util/filterDuplicates.js
