+++ server/OCR/util/filterOverlapping.js
