+++ server/OCR/util/getCenterVertex.js
