+++ server/OCR/util/getDetectionDimensions.js
