+++ server/OCR/util/getDimensions.js
