+++ server/OCR/util/getHand.js
