+++ server/OCR/util/getMinMaxVertices.js
