+++ server/OCR/util/pixelToPercent.js
