+++ server/algorithm/data/board.js
