+++ server/algorithm/data/cachePermutations.js
