+++ server/algorithm/data/emptyState.js
