+++ server/algorithm/data/findNewWords.js
