+++ server/algorithm/data/letterScores.js
