+++ server/algorithm/data/permutations.js
