+++ server/algorithm/data/score.js
