+++ server/algorithm/data/trie.js
