+++ server/algorithm/permAlgorithm.js
