+++ server/package.json
@@ -13,9 +13,11 @@
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
+		"@google-cloud/vision": "^5.1.0",
 		"dotenv": "^16.4.7",
 		"express": "^4.21.2",
-		"multer": "^1.4.5-lts.1"
+		"multer": "^1.4.5-lts.1",
+		"word-list": "^4.1.0"
 	},
 	"devDependencies": {
 		"nodemon": "^3.1.9"
